requirements/requirements.txt
hydra-core
matplotlib
numpy
pandas
pywavelets
scipy
torch

